+++ app/src/main/java/ajou/com/skechip/Retrofit/models/MeetingResponse.java
@@ -7,14 +7,14 @@ public class MeetingResponse {
 
     private List<Integer> idList;
     private List<Integer> typeList;
-    private List<Integer> managerList;
+    private List<Long> managerList;
     private List<String> titleList;
     private List<String> placeList;
     private List<Integer> cellPositionList;
     private  Integer totalCount;
     private Integer groupId;
 
-    public MeetingResponse(boolean error, List<Integer> idList, List<Integer> typeList, List<Integer> managerList, List<String> titleList, List<String> placeList, List<Integer> cellPositionList, Integer totalCount, Integer groupId) {
+    public MeetingResponse(boolean error, List<Integer> idList, List<Integer> typeList, List<Long> managerList, List<String> titleList, List<String> placeList, List<Integer> cellPositionList, Integer totalCount, Integer groupId) {
         this.error = error;
         this.idList = idList;
         this.typeList = typeList;
@@ -38,7 +38,7 @@ public class MeetingResponse {
         return typeList;
     }
 
-    public List<Integer> getManagerList() {
+    public List<Long> getManagerList() {
         return managerList;
     }
